Audit, fix, and expand the notebook; rebuild it for print
Nine research agents audited every section against KACTL, ShahjalalShohag/code-library,
cp-algorithms, AtCoder Library, OI Wiki, Library Checker and real ECPC/ACPC/Codeforces sets.
Every bug below was reproduced locally (compiled, brute-forced, or run under UBSan) before the
fix went in, and every added routine is stress-tested against an independent reference.
25 verification suites, 0 failures.

Crashes and hangs fixed
- PersistentLiChaoTree: (l+r)/2 truncates toward zero, so on a negative interval the left child
  equals the parent - infinite recursion on ~2/3 of negative query abscissae.
- minkowski: on clockwise input neither merge index advances - an unexplained TLE/MLE.
- Offline2DFenwickTree: constructor parameter shadowed the member, so add() walked one row past
  the end of vals.
- Sieve1e9: inner loop bound n/2 wrote one index past a 60 MB bitset.
- Empty-input UB in mec, inConvex, KD::nearest, closestPair, TreeIsomorphism::centers,
  DynamicMedian::remove_val, BerlekampMassey::linearRec.

Silent wrong answers fixed
- patternRange never reset R, so an absent pattern returned a non-empty suffix-array range.
- wildcardMatch used c - 'a' + 1, negative below 'a', and the FFT identity needs non-negative
  values; terms cancelled and reported matches that do not exist. Now compresses the alphabet.
- IterativeSegTree::update reversed its arguments for every right child - wrong for any
  non-commutative merge.
- CHT compared ~2e18 products in long double; LiChaoTree computed the crossing abscissa in
  double. Both now exact (__int128 / integer comparisons at l, mid, r).
- SegmentTreeBeats held every value in int; Hash2D re-seeded its bases per object, so two grids
  were never comparable.
- staircaseLosing's code and comment used opposite ground conventions; Periodic::detect took the
  first matching index instead of the last; lasker(0) returned -1; pathsAvoidingLine and ballot
  returned negative counts; perpBisector on P<ll> returned a different line.
- segCross inherited segInter's division, which overflows above |c| ~ 1.6e6; it is now sign-only.
- sgn(int) and sgn(double) were ambiguous - a hard compile error on any non-ll argument.
- GaussianElimination dropped every free column after the last pivot.
- NTT silently returns garbage above length 2^23; asserted and documented.

Preconditions corrected
KnuthDP and AliensTrick documented their conclusions instead of their checkable hypotheses;
qBinom's real precondition is ord_p(q) > k; Anti-SG needs SG(u)=0 iff u terminal; Hall needs
"subset of the left part"; Dilworth needs "finite poset". MoOnPaths could not be called at all -
the data needed to build its queries was private and computed only inside solve().

Added
- New section 12 Problem Solving: reframing, search, proof tools, constraints-as-hints,
  constructive and interactive, randomisation, empirics, contest craft, pre-submit checklist,
  verdict triage.
- Graph theorem sheet, min-cut modelling card, exact-exponential graph algorithms (Bron-Kerbosch,
  chromatic number, Steiner tree), triangle and 4-cycle counting, second-best MST.
- Tree DP card, Euler-tour toolbox, mergeable heap, monotone CHT.
- Second number-theory formula sheet, floor_sum, adaptive Simpson, Jacobi symbol.
- Nimbers with the octal-game Grundy generator, anti-SG, every-SG, Euclid's game.
- Wildcard matching, generalized SAM, Josephus, BigInt.
- Geometry: polygon triangulation, polygon operations (isSimple, winding number, clipping,
  convex distance, tangents, line-hull) and the whole small-primitive layer (angle bisectors,
  barycentric coordinates, in-circle predicate, Apollonius, max points covered, geometric median).

Rebuilt for print
Alphabetical index of every routine (520 entries), running head naming section and sub-topic, a
one-line description under every file title, a HOW TO FIND SOMETHING box on page 1, denser TOC,
no listing line numbers, and orphan control on titles.

docs/FROM-OLD-TEMPLATES.md records what the temps/ reference libraries still have that this
notebook does not, as a prioritised worklist.

+++ notebook/02 - Data Structures (DS)/01 - Segment Tree/03 - PersistentSegmentTree.cpp
@@ -1,8 +1,11 @@
+// PERSISTENT SEGMENT TREE - every update returns a NEW root and old versions stay queryable.
+// O(log n) time and O(log n) new nodes per update. The standard use is offline k-th smallest in a
+// range (build a version per prefix, then query version[r] minus version[l-1]).
 struct Node {
     Node *l, *r;
     ll val = 0;
 
-    Node(int val) : l(NULL), r(NULL), val(val) {
+    Node(ll val) : l(NULL), r(NULL), val(val) {
     }
 
     Node() : l(NULL), r(NULL) {
@@ -31,7 +34,7 @@ struct PST {
         return ret;
     }
 
-    Node* update(Node* v, int i, int val, int lx, int rx) {
+    Node* update(Node* v, int i, ll val, int lx, int rx) {
         if (lx == rx) return new Node(v->val + val);
         int mid = (lx + rx) / 2;
         if (!v->l) v->addChild();
@@ -39,7 +42,7 @@ struct PST {
         return new Node(v->l, update(v->r, i, val, mid + 1, rx));
     }
 
-    Node* update(Node* v, int i, int val) { return update(v, i, val, 0, n - 1); }
+    Node* update(Node* v, int i, ll val) { return update(v, i, val, 0, n - 1); }
 
     Node query(Node* v, int l, int r, int lx, int rx) {
         if (l > rx || r < lx) return {};
